refacto(folios) refactor update occurence card in main folio
