+++ app/controllers/product/card_folios_controller.ts
@@ -5,7 +5,7 @@ import { errors as vineErrors } from '@vinejs/vine'
 import NotFoundException from '#exceptions/not_found_exception'
 import FolioService from '#services/folio_service'
 import { SuccessOutputDTO } from '#types/success_output_dto_type'
-import { occurrenceValidator, removeMainValidator } from '#validators/card_folio_validator'
+import { removeMainValidator } from '#validators/card_folio_validator'
 import CardService from '#services/card_service'
 import CardFolioService from '#services/card_folio_service'
 import ValidationException from '#exceptions/validation_exception'
@@ -17,27 +17,6 @@ export default class CardFoliosController {
     private cardService: CardService,
     private cardFolioService: CardFolioService
   ) {}
-
-  async occurrence({ request, response, authUser }: HttpContext) {
-    try {
-      const payload = await request.validateUsing(occurrenceValidator)
-      const card = await this.cardService.getCardIdById(payload.params.id)
-      const folio = await this.folioService.getMainFolioByUserId(authUser.id)
-      await this.cardFolioService.updateCardFolioOccurrence(card.id, folio.id, payload.occurrence)
-      const successResponse: SuccessOutputDTO = {
-        message: 'Card occurrence updated successfully',
-      }
-      return response.ok(successResponse)
-    } catch (error) {
-      if (error instanceof vineErrors.E_VALIDATION_ERROR) {
-        throw new ValidationException(error)
-      }
-      if (error instanceof lucidErrors.E_ROW_NOT_FOUND) {
-        throw new NotFoundException(error)
-      }
-      throw error
-    }
-  }
 
   async delete({ request, response, authUser }: HttpContext) {
     try {
