+++ app/controllers/product/main_folios_controller.ts
@@ -12,7 +12,7 @@ import CardFolioMapper from '#mappers/card_folio_mapper'
 import { SuccessOutputDTO } from '#types/success_output_dto_type'
 import CardService from '#services/card_service'
 import DuplicateEntryException from '#exceptions/duplicate_entry_exception'
-import { collectValidator } from '#validators/folio_validator'
+import { collectValidator, updateOccurrenceValidator } from '#validators/folio_validator'
 
 @inject()
 export default class MainFoliosController {
@@ -99,6 +99,27 @@ export default class MainFoliosController {
       }
       if (error.code === 'ER_DUP_ENTRY') {
         throw new DuplicateEntryException('Card already exists in the folio')
+      }
+      if (error instanceof lucidErrors.E_ROW_NOT_FOUND) {
+        throw new NotFoundException(error)
+      }
+      throw error
+    }
+  }
+
+  async updateOccurrence({ request, response, authUser }: HttpContext) {
+    try {
+      const payload = await request.validateUsing(updateOccurrenceValidator)
+      const card = await this.cardService.getCardIdById(payload.params.id)
+      const folio = await this.folioService.getMainFolioByUserId(authUser.id)
+      await this.cardFolioService.updateCardFolioOccurrence(card.id, folio.id, payload.occurrence)
+      const successResponse: SuccessOutputDTO = {
+        message: 'Card occurrence updated successfully',
+      }
+      return response.ok(successResponse)
+    } catch (error) {
+      if (error instanceof vineErrors.E_VALIDATION_ERROR) {
+        throw new ValidationException(error)
       }
       if (error instanceof lucidErrors.E_ROW_NOT_FOUND) {
         throw new NotFoundException(error)
